Modifico tipo int por double en parámetro de poderPelea

+++ src/modelo/personajes/modos/ModoNormal.java
@@ -4,7 +4,7 @@ import modelo.personajes.Personaje;
 
 public class ModoNormal extends Modo{
 
-	public ModoNormal(int poderPelea, int distanciaAtaque, int velocidad, String nombre) {
+	public ModoNormal(double poderPelea, int distanciaAtaque, int velocidad, String nombre) {
 		super(poderPelea, distanciaAtaque, velocidad, nombre);
 	}
