+++ src/modelo/personajes/modos/PrimeraTransformacion.java
@@ -5,7 +5,7 @@ import modelo.personajes.Personaje;
 public class PrimeraTransformacion extends Modo {
 
 	
-	public PrimeraTransformacion(int poderPelea, int distanciaAtaque, int velocidad, String nombre) {
+	public PrimeraTransformacion(double poderPelea, int distanciaAtaque, int velocidad, String nombre) {
 		super(poderPelea, distanciaAtaque, velocidad, nombre);
 	}
